Treat INADDR_NONE as uint32_t.
Reviewed by:	glebius


Former-commit-id: 621dcbbd6a58ef99a1702b9e1267311cdec21822

+++ sys/netinet/in.h
@@ -2095,7 +2095,7 @@ begin_define
 define|#
 directive|define
 name|INADDR_NONE
-value|0xffffffff
+value|(uint32_t)0xffffffff
 end_define
 
 begin_comment
